Rename DatabaseAdapter to Database

+++ src/main/java/org/riverock/dbrevision/db/factory/DB2Database.java
@@ -25,6 +25,9 @@
  */
 package org.riverock.dbrevision.db.factory;
 
+import java.io.ByteArrayOutputStream;
+import java.io.InputStream;
+import java.sql.Blob;
 import java.sql.Connection;
 import java.sql.DatabaseMetaData;
 import java.sql.PreparedStatement;
@@ -34,54 +37,52 @@ import java.sql.Statement;
 import java.sql.Types;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Collections;
 
-import org.apache.commons.lang.StringUtils;
 import org.apache.log4j.Logger;
 
 import org.riverock.dbrevision.annotation.schema.db.DbDataFieldData;
 import org.riverock.dbrevision.annotation.schema.db.DbField;
 import org.riverock.dbrevision.annotation.schema.db.DbForeignKey;
 import org.riverock.dbrevision.annotation.schema.db.DbPrimaryKey;
+import org.riverock.dbrevision.annotation.schema.db.DbPrimaryKeyColumn;
 import org.riverock.dbrevision.annotation.schema.db.DbSequence;
 import org.riverock.dbrevision.annotation.schema.db.DbTable;
 import org.riverock.dbrevision.annotation.schema.db.DbView;
-import org.riverock.dbrevision.annotation.schema.db.DbPrimaryKeyColumn;
-import org.riverock.dbrevision.db.DatabaseAdapter;
+import org.riverock.dbrevision.db.Database;
 import org.riverock.dbrevision.db.DatabaseManager;
-import org.riverock.dbrevision.db.DbPkComparator;
 import org.riverock.dbrevision.exception.DbRevisionException;
 import org.riverock.dbrevision.utils.DbUtils;
+import org.riverock.dbrevision.utils.Utils;
 
 /**
- * InterBase database connect 
+ * IBM DB2 connection
  * $Author: serg_main $
  *
- * $Id: InterbaseAdapter.java 1141 2006-12-14 14:43:29Z serg_main $
+ * $Id: DB2Database.java 1141 2006-12-14 14:43:29Z serg_main $
  *
  */
 @SuppressWarnings({"UnusedAssignment"})
-public class InterbaseAdapter extends DatabaseAdapter {
-    private static Logger log = Logger.getLogger( InterbaseAdapter.class );
+public class DB2Database extends Database {
+    private static Logger log = Logger.getLogger(DB2Database.class);
 
     /**
      * get family for this adapter
      * @return family
      */
     public Family getFamily() {
-        return Family.INTERBASE;
+        return Family.DB2;
     }
 
-    public InterbaseAdapter(Connection conn) {
+    public DB2Database(Connection conn) {
         super(conn);
     }
 
     public int getMaxLengthStringField() {
-        return 4000;
+        return 2000;
     }
 
     public boolean isBatchUpdate() {
-        return false;
+        return true;
     }
 
     public boolean isNeedUpdateBracket() {
@@ -97,15 +98,31 @@ public class InterbaseAdapter extends DatabaseAdapter {
     }
 
     public byte[] getBlobField(ResultSet rs, String nameField, int maxLength) {
-        return null;
+        try {
+            Blob blob = rs.getBlob(nameField);
+            ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
+            int count;
+            byte buffer[] = new byte[1024];
+
+            InputStream inputStream = blob.getBinaryStream();
+            while ((count = inputStream.read(buffer)) >= 0) {
+                outputStream.write(buffer, 0, count);
+                outputStream.flush();
+            }
+            outputStream.close();
+            return outputStream.toByteArray();
+        }
+        catch (Exception e) {
+            throw new DbRevisionException(e);
+        }
     }
 
     public void createTable(DbTable table) {
-        if (table == null || table.getFields().isEmpty() ) {
+        if (table == null || table.getFields().isEmpty()) {
             return;
         }
 
-        String sql = "create table " + table.getName() + "\n" +
+        String sql = "create table \"" + table.getName() + "\" " +
             "(";
 
         boolean isFirst = true;
@@ -116,20 +133,12 @@ public class InterbaseAdapter extends DatabaseAdapter {
             else
                 isFirst = !isFirst;
 
-            sql += "\n" + field.getName();
+            sql += " \"" + field.getName() + "\"";
             switch (field.getJavaType()) {
 
                 case Types.NUMERIC:
                 case Types.DECIMAL:
-                    if (field.getDecimalDigit()!=0) {
-                        sql += " DECIMAL(" + (field.getSize()==null || field.getSize() > 38 ? 38 : field.getSize()) + ',' + field.getDecimalDigit() + ")";
-                    }
-                    else {
-                        if (field.getSize() == 1)
-                            sql += " SMALLINT";
-                        else
-                            sql += " DOUBLE PRECISION";
-                    }
+                    sql += " DECIMAL(" + (field.getSize()==null || field.getSize() > 31 ? 31 : field.getSize()) + ',' + field.getDecimalDigit() + ")";
                     break;
 
                 case Types.INTEGER:
@@ -150,16 +159,14 @@ public class InterbaseAdapter extends DatabaseAdapter {
 
                 case Types.TIMESTAMP:
                 case Types.DATE:
-                    sql += " DATETIME";
+                    sql += " TIMESTAMP";
                     break;
 
                 case Types.LONGVARCHAR:
-                    // Oracle 'long' fields type
                     sql += " VARCHAR(10)";
                     break;
 
                 case Types.LONGVARBINARY:
-                    // Oracle 'long raw' fields type
                     sql += " LONGVARBINARY";
                     break;
 
@@ -171,60 +178,61 @@ public class InterbaseAdapter extends DatabaseAdapter {
             if (field.getDefaultValue() != null) {
                 String val = field.getDefaultValue().trim();
 
-                // TODO rewrite. check only if type is 'date' 
-                if (DatabaseManager.checkDefaultTimestamp(val)) {
-                    val = "current_timestamp";
-                }
+//                if (!val.equalsIgnoreCase("null"))
+//                    val = "'"+val+"'";
+                if (DatabaseManager.checkDefaultTimestamp(val))
+                    val = "CURRENT TIMESTAMP";
 
                 sql += (" DEFAULT " + val);
             }
 
-            if (field.getNullable()==DatabaseMetaData.columnNoNulls) {
+            if (field.getNullable() == DatabaseMetaData.columnNoNulls) {
                 sql += " NOT NULL ";
             }
         }
-        if (table.getPrimaryKey() != null && table.getPrimaryKey().getColumns().size() != 0) {
+
+        if (table.getPrimaryKey() != null && table.getPrimaryKey().getColumns().size() > 0) {
             DbPrimaryKey pk = table.getPrimaryKey();
 
-            //            constraintDefinition:
-//            [ CONSTRAINT name ]
-//            UNIQUE ( column [,column...] ) |
-//            PRIMARY KEY ( column [,column...] ) |
+            sql += ", CONSTRAINT " + pk.getPkName() + " PRIMARY KEY ( ";
 
-            sql += ",\nCONSTRAINT " + pk.getPkName() + " PRIMARY KEY (\n";
-
-            Collections.sort(pk.getColumns(), DbPkComparator.getInstance());
+            int seq = Integer.MIN_VALUE;
             isFirst = true;
-            for (DbPrimaryKeyColumn column : pk.getColumns()) {
-                if (!isFirst) {
+            for (DbPrimaryKeyColumn primaryKeyColumnType : pk.getColumns()) {
+                DbPrimaryKeyColumn column = primaryKeyColumnType;
+                int seqTemp = Integer.MAX_VALUE;
+                for (DbPrimaryKeyColumn columnTemp : pk.getColumns()) {
+                    if (seq < columnTemp.getKeySeq() && columnTemp.getKeySeq() < seqTemp) {
+                        seqTemp = columnTemp.getKeySeq();
+                        column = columnTemp;
+                    }
+                }
+                seq = column.getKeySeq();
+
+                if (!isFirst)
                     sql += ",";
-                }
-                else {
+                else
                     isFirst = !isFirst;
-                }
 
                 sql += column.getColumnName();
             }
-            sql += "\n)";
+            sql += " )";
         }
-        sql += "\n)";
 
-        Statement st = null;
+        sql += " )";
+
+        Statement ps = null;
         try {
-            st = this.getConnection().createStatement();
-            st.execute(sql);
-            int count = st.getUpdateCount();
-            if (log.isDebugEnabled()) {
-                log.debug("count of processed records " + count);
-            }
+            ps = this.getConnection().createStatement();
+            ps.executeUpdate(sql);
+            this.getConnection().commit();
         }
         catch (SQLException e) {
-            log.error("SQL:\n"+sql);
             throw new DbRevisionException(e);
         }
         finally {
-            DbUtils.close(st);
-            st = null;
+            DbUtils.close(ps);
+            ps = null;
         }
 
     }
@@ -233,31 +241,29 @@ public class InterbaseAdapter extends DatabaseAdapter {
     }
 
     public void dropTable(DbTable table) {
+        if (table == null)
+            return;
+
         dropTable(table.getName());
     }
 
     public void dropTable(String nameTable) {
-        if (nameTable == null) {
+        if (nameTable == null || nameTable.trim().length() == 0)
             return;
-        }
 
         String sql = "drop table " + nameTable;
 
-        Statement st = null;
+        Statement ps = null;
         try {
-            st = this.getConnection().createStatement();
-            st.execute(sql);
-            int count = st.getUpdateCount();
-            if (log.isDebugEnabled())
-                log.debug("count of deleted object " + count);
+            ps = this.getConnection().createStatement();
+            ps.executeUpdate(sql);
         }
         catch (SQLException e) {
-            log.error("Error drop table " + nameTable, e);
             throw new DbRevisionException(e);
         }
         finally {
-            DbUtils.close(st);
-            st = null;
+            DbUtils.close(ps);
+            ps = null;
         }
     }
 
@@ -286,164 +292,15 @@ public class InterbaseAdapter extends DatabaseAdapter {
     }
 
     public void addColumn(DbTable table, DbField field) {
-        String sql = "alter table " + table.getName() + " add " + field.getName() + " ";
-
-        switch (field.getJavaType()) {
-
-            case Types.NUMERIC:
-            case Types.DECIMAL:
-                sql += " DOUBLE PRECISION";
-                break;
-
-            case Types.INTEGER:
-                sql += " INTEGER";
-                break;
-
-            case Types.DOUBLE:
-                sql += " DOUBLE";
-                break;
-
-            case Types.CHAR:
-                sql += " VARCHAR(1)";
-                break;
-
-            case Types.VARCHAR:
-                sql += (" VARCHAR(" + field.getSize() + ") ");
-                break;
-
-            case Types.TIMESTAMP:
-            case Types.DATE:
-                sql += " DATETIME";
-                break;
-
-            case Types.LONGVARCHAR:
-                // Oracle 'long' fields type
-                sql += " VARCHAR(10)";
-                break;
-
-            case Types.LONGVARBINARY:
-                // Oracle 'long raw' fields type
-                sql += " LONGVARBINARY";
-                break;
-
-            default:
-                field.setJavaStringType("unknown field type field - " + field.getName() + " javaType - " + field.getJavaType());
-                System.out.println("unknown field type field - " + field.getName() + " javaType - " + field.getJavaType());
-        }
-
-        if (field.getDefaultValue() != null) {
-            String val = field.getDefaultValue().trim();
-
-            //TODO rewrite init of def as in createTable
-//                if (!val.equalsIgnoreCase("null"))
-//                    val = "'"+val+"'";
-            if (DatabaseManager.checkDefaultTimestamp(val))
-                val = "current_timestamp";
-
-            sql += (" DEFAULT " + val);
-        }
-
-        if (field.getNullable() == DatabaseMetaData.columnNoNulls) {
-            sql += " NOT NULL ";
-        }
-
-        if (log.isDebugEnabled())
-            log.debug("Interbase addColumn sql - \n" + sql);
-
-        Statement ps = null;
-        try {
-            ps = this.getConnection().createStatement();
-            ps.executeUpdate(sql);
-            this.getConnection().commit();
-        }
-        catch (SQLException e) {
-            throw new DbRevisionException(e);
-        }
-        finally {
-            DbUtils.close(ps);
-            ps = null;
-        }
     }
 
     public String getOnDeleteSetNull() {
-        return "ON DELETE NO ACTION";
+        return null;
     }
 
     public String getDefaultTimestampValue() {
-        return "current_timestamp";
+        return "CURRENT TIMESTAMP";
     }
-
-/*
-ALTER TABLE table
-{ [ ALTER COLUMN column_name
-    { new_data_type [ ( precision [ , scale ] ) ]
-        [ COLLATE < collation_name > ]
-        [ NULL | NOT NULL ]
-        | {ADD | DROP } ROWGUIDCOL }
-    ]
-    | ADD
-        { [ < column_definition > ]
-        |  column_name AS computed_column_expression
-        } [ ,...n ]
-    | [ WITH CHECK | WITH NOCHECK ] ADD
-        { < table_constraint > } [ ,...n ]
-    | DROP
-        { [ CONSTRAINT ] constraint_name
-            | COLUMN column } [ ,...n ]
-    | { CHECK | NOCHECK } CONSTRAINT
-        { ALL | constraint_name [ ,...n ] }
-    | { ENABLE | DISABLE } TRIGGER
-        { ALL | trigger_name [ ,...n ] }
-}
-
-< column_definition > ::=
-    { column_name data_type }
-    [ [ DEFAULT constant_expression ] [ WITH VALUES ]
-    | [ IDENTITY [ (seed , increment ) [ NOT FOR REPLICATION ] ] ]
-        ]
-    [ ROWGUIDCOL ]
-    [ COLLATE < collation_name > ]
-    [ < column_constraint > ] [ ...n ]
-
-< column_constraint > ::=
-    [ CONSTRAINT constraint_name ]
-    { [ NULL | NOT NULL ]
-        | [ { PRIMARY KEY | UNIQUE }
-            [ CLUSTERED | NONCLUSTERED ]
-            [ WITH FILLFACTOR = fillfactor ]
-            [ ON { filegroup | DEFAULT } ]
-            ]
-        | [ [ FOREIGN KEY ]
-            REFERENCES ref_table [ ( ref_column ) ]
-            [ ON DELETE { CASCADE | NO ACTION } ]
-            [ ON UPDATE { CASCADE | NO ACTION } ]
-            [ NOT FOR REPLICATION ]
-            ]
-        | CHECK [ NOT FOR REPLICATION ]
-            ( logical_expression )
-    }
-
-< table_constraint > ::=
-    [ CONSTRAINT constraint_name ]
-    { [ { PRIMARY KEY | UNIQUE }
-        [ CLUSTERED | NONCLUSTERED ]
-        { ( column [ ,...n ] ) }
-        [ WITH FILLFACTOR = fillfactor ]
-        [ ON {filegroup | DEFAULT } ]
-        ]
-        |    FOREIGN KEY
-            [ ( column [ ,...n ] ) ]
-            REFERENCES ref_table [ ( ref_column [ ,...n ] ) ]
-            [ ON DELETE { CASCADE | NO ACTION } ]
-            [ ON UPDATE { CASCADE | NO ACTION } ]
-            [ NOT FOR REPLICATION ]
-        | DEFAULT constant_expression
-            [ FOR column ] [ WITH VALUES ]
-        |    CHECK [ NOT FOR REPLICATION ]
-            ( search_conditions )
-    }
-
-*/
 
     public List<DbView> getViewList(String schemaPattern, String tablePattern) {
         return DatabaseManager.getViewList(getConnection(), schemaPattern, tablePattern);
@@ -466,7 +323,9 @@ ALTER TABLE table
 
         String sql_ =
             "CREATE VIEW " + view.getName() +
-            " AS " + StringUtils.replace(view.getText(), "||", "+");
+            " AS " +
+            Utils.replaceStringArray(view.getText(),
+                new String[][]{{"||", "+"}, {"\n", " "}}).trim();
 
         Statement ps = null;
         try {
@@ -519,8 +378,10 @@ ALTER TABLE table
 */
 
     public boolean testExceptionTableNotFound(Exception e) {
-        if (((SQLException) e).getErrorCode() == 208)
-            return true;
+        if (e instanceof SQLException) {
+            if (((SQLException) e).getErrorCode() == -204)
+                return true;
+        }
         return false;
     }
 
@@ -545,7 +406,7 @@ ALTER TABLE table
 
     public boolean testExceptionTableExists(Exception e) {
         if (e instanceof SQLException) {
-            if (((SQLException) e).getErrorCode() == 335544351)
+            if (((SQLException) e).getErrorCode() == -601)
                 return true;
         }
         return false;
@@ -553,7 +414,7 @@ ALTER TABLE table
 
     public boolean testExceptionViewExists(Exception e) {
         if (e instanceof SQLException) {
-            if (((SQLException) e).getErrorCode() == 2714)
+            if (((SQLException) e).getErrorCode() == -601)
                 return true;
         }
         return false;
